server_requirements.txt
anyio
mcp
requests
pyyaml
Pillow
python-dotenv
